User & Company page back working after changed to new api

+++ 4thYearProjectAPI/4thYearProjectAPI/Models/Company.cs
@@ -9,7 +9,7 @@ namespace _4thYearProjectAPI.Models
     public class Company
     {
         [Key]
-        public double compID { get; set; }
+        public int compID { get; set; }
         public string compName { get; set; }
     }
 }
